+++ src/main/js/lightbox.js
@@ -27,12 +27,42 @@ class Lightbox {
     $target.addEventListener('click', e => {
       $target.close();
     });
+
+    // Keyboard navigation
     $target.addEventListener('keydown', e => {
       let offset;
       switch (e.key) {
         case 'ArrowLeft': offset = parseInt(e.target.dataset.offset) - 1; break;
         case 'ArrowRight': offset = parseInt(e.target.dataset.offset) + 1; break;
         default: return;
+      }
+
+      e.stopPropagation();
+      selector.item(offset)?.dispatchEvent(new Event('click'));
+    });
+
+    // Swipe left and right
+    let x, y;
+    const threshold = 50;
+    $target.addEventListener('touchstart', e => {
+      x = e.touches[0].clientX;
+      y = e.touches[0].clientY;
+    });
+    $target.addEventListener('touchmove', e => e.cancelable && e.preventDefault(), { passive: false });
+    $target.addEventListener('touchend', e => {
+      const width = e.changedTouches[0].clientX - x;
+      const height = e.changedTouches[0].clientY - y;
+
+      // Swipe was mostly vertical, ignore
+      if (Math.abs(width) <= Math.abs(height)) return;
+
+      let offset;
+      if (width > threshold) {
+        offset = parseInt($target.dataset.offset) - 1;
+      } else if (width < -threshold) {
+        offset = parseInt($target.dataset.offset) + 1;
+      } else {
+        return;
       }
 
       e.stopPropagation();